requirements.txt
playwright
openai
python-dotenv
dateparser
python-dateutil
discord.py
